requirements/pyproject.toml
[tool.poetry]
name = "file-manipulator"
version = "0.1.0"
description = "An util for file manipulation"
authors = ["Darya Litvinchuk"]

[tool.poetry.dependencies]
python = "~3.8.0"

python-dotenv = "~0.17.1"
pydantic = "~1.8.2"

mypy = "~0.902"
flake8 = "~3.9.2"
flake8_polyfill = "~1.0.2"
isort = "~5.8.0"
black = "~21.6b0"
pytest = "~6.2.4"
pytest-cov = "~2.12.1"
mock = "~4.0.3"

radon = "~5.0.1"
mutmut = "~2.1.0"

pip-licenses = "~3.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
